+++ app/src/main/java/com/example/movopfy/app/App.kt
@@ -1,8 +1,8 @@
 package com.example.movopfy.app
 
 import android.app.Application
+import com.example.movopfy.features.details.koin.detailsModule
 import com.example.movopfy.features.home.koin.homeModule
-import com.example.movopfy.features.title.koin.titleModule
 import com.example.movopfy.network.koin.networkModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
@@ -17,7 +17,7 @@ class App : Application() {
         startKoin {
             androidLogger(Level.DEBUG)
             androidContext(this@App)
-            modules(listOf(homeModule, networkModule, titleModule))
+            modules(listOf(homeModule, networkModule, detailsModule))
         }
     }
 }
