+++ app/src/main/java/com/example/movopfy/features/details/data/repository/AnilibriaRepositoryImpl.kt
@@ -1,6 +1,6 @@
-package com.example.movopfy.features.title.data.repository
+package com.example.movopfy.features.details.data.repository
 
-import com.example.movopfy.features.title.domain.repository.AnilibriaRepository
+import com.example.movopfy.features.details.domain.repository.AnilibriaRepository
 import com.example.movopfy.network.anilibria.models.AnilibriaTitle
 import com.example.movopfy.network.anilibria.service.AnilibriaService
 import kotlinx.coroutines.Dispatchers
