+++ app/src/main/java/com/example/movopfy/features/details/presentation/ui/DescriptionText.kt
@@ -1,4 +1,4 @@
-package com.example.movopfy.features.title.presentation.ui
+package com.example.movopfy.features.details.presentation.ui
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
@@ -6,17 +6,19 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.example.movopfy.R
 import com.example.movopfy.uiComponents.theme.Dimensions
 import com.example.movopfy.uiComponents.theme.TextMain
 
 @Composable
-fun NameOfTitle(dimensions: Dimensions, name: String?) {
+fun DescriptionText(dimensions: Dimensions, description: String?) {
     Text(
-        modifier = Modifier.padding(start = dimensions.paddingStart),
-        text = name.toString(),
-        fontFamily = FontFamily(Font(R.font.inter_extrabold)),
-        fontSize = dimensions.textSizeMain,
+        text = description.toString(),
+        modifier = Modifier.padding(start = dimensions.paddingStart, top = 20.dp),
+        fontFamily = FontFamily(Font(R.font.inter_bold)),
+        fontSize = 15.sp,
         color = TextMain
     )
 }
