+++ app/src/main/java/com/example/movopfy/features/details/presentation/viewmodel/DetailsViewModel.kt
@@ -1,15 +1,15 @@
-package com.example.movopfy.features.title.presentation.viewmodel
+package com.example.movopfy.features.details.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.movopfy.features.title.domain.repository.AnilibriaRepository
+import com.example.movopfy.features.details.domain.repository.AnilibriaRepository
 import com.example.movopfy.network.anilibria.models.AnilibriaTitle
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-class TitleViewModel(private val anilibriaRepository: AnilibriaRepository) : ViewModel() {
+class DetailsViewModel(private val anilibriaRepository: AnilibriaRepository) : ViewModel() {
 
     private val _uiState = MutableStateFlow<TitleUiState>(TitleUiState.Loading)
     val uiState: StateFlow<TitleUiState> = _uiState.asStateFlow()
