+++ app/src/main/java/com/example/movopfy/features/home/presentation/ui/SchedulesHeader.kt
@@ -1,24 +1,27 @@
-package com.example.movopfy.features.title.presentation.ui
+package com.example.movopfy.features.home.presentation.ui
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.example.movopfy.R
 import com.example.movopfy.uiComponents.theme.Dimensions
 import com.example.movopfy.uiComponents.theme.TextMain
 
 @Composable
-fun DescriptionOfTitle(dimensions: Dimensions, description: String?) {
+fun SchedulesHeader(dimensions: Dimensions) {
     Text(
-        text = description.toString(),
-        modifier = Modifier.padding(start = dimensions.paddingStart, top = 20.dp),
+        text = stringResource(R.string.schedule_for_today_text),
+        color = TextMain,
+        fontSize = dimensions.textSizeMain,
         fontFamily = FontFamily(Font(R.font.inter_bold)),
-        fontSize = 15.sp,
-        color = TextMain
+        modifier = Modifier.padding(
+            start = dimensions.paddingStart,
+            top = 100.dp
+        )
     )
 }
