+++ app/src/main/java/com/example/movopfy/uiComponents/navigation/Screen.kt
@@ -1,14 +1,14 @@
 package com.example.movopfy.uiComponents.navigation
 
-const val TITLE_ID = "id"
+const val DETAILS_ID = "id"
 
 sealed class Screen(val route: String) {
 
     data object Home : Screen(route = "home_screen")
 
-    data object Title : Screen(route = "title_screen/{$TITLE_ID}") {
+    data object Details : Screen(route = "details_screen/{$DETAILS_ID}") {
         fun passId(id: Int): String {
-            return this.route.replace(oldValue = "{$TITLE_ID}", newValue = id.toString())
+            return this.route.replace(oldValue = "{$DETAILS_ID}", newValue = id.toString())
         }
     }
 }
